Nav-bar 검색 기능 구현 및 반응형 디자인
