+++ src/index.css
@@ -73,6 +73,6 @@ button:focus-visible {
 
 body {
   background-color: #111;
-  color: white;
+  color: #eee;
   font-family: sans-serif;
 }
